+++ packages/components-angular/src/app/pages/checkbox-wrapper.component.ts
@@ -3,6 +3,21 @@ import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
 
 @Component({
   selector: 'page-checkbox-wrapper',
+  styles: [
+    `
+      @media only screen and (min-width: 760px) {
+        #app,
+        :host {
+          display: grid;
+          grid-template-columns: repeat(2, 50%);
+        }
+      }
+    
+      .playground > * {
+        margin-bottom: 8px;
+      }
+    `,
+  ],
   template: `
     <div class="playground light" title="should render with label">
       <p-checkbox-wrapper [label]="'Some label'">
@@ -45,6 +60,23 @@ import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
         <input [type]="'checkbox'" [name]="'some-name'" />
       </p-checkbox-wrapper>
       <p-checkbox-wrapper [label]="'Some label'" [hideLabel]="{ base: true, xs: false, s: true, m: false, l: true, xl: false }">
+        <input [type]="'checkbox'" [name]="'some-name'" checked />
+      </p-checkbox-wrapper>
+    </div>
+
+    <div class="playground dark" title="should render with responsive label">
+      <p-checkbox-wrapper
+        [theme]="'dark'"
+        [label]="'Some label'"
+        [hideLabel]="{ base: true, xs: false, s: true, m: false, l: true, xl: false }"
+      >
+        <input [type]="'checkbox'" [name]="'some-name'" />
+      </p-checkbox-wrapper>
+      <p-checkbox-wrapper
+        [theme]="'dark'"
+        [label]="'Some label'"
+        [hideLabel]="{ base: true, xs: false, s: true, m: false, l: true, xl: false }"
+      >
         <input [type]="'checkbox'" [name]="'some-name'" checked />
       </p-checkbox-wrapper>
     </div>
@@ -195,17 +227,73 @@ import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
 
     <div class="playground light" title="should render with slotted content with error state and message">
       <p-checkbox-wrapper [state]="'error'">
-        <span slot="label">Some label with a <a [href]="'https://designsystem.porsche.com'">link</a>.</span>
+        <span slot="label">
+          <span>
+            Slotted label. And some slotted and deeply nested <a [href]="'#'">anchor</a>, <b>bold</b>, <strong>strong</strong>,
+            <em>emphasized</em> and <i>italic</i> text.
+          </span>
+        </span>
         <input [type]="'checkbox'" [name]="'some-name'" />
-        <span slot="message">Some error message with a <a [href]="'https://designsystem.porsche.com'">link</a>.</span>
+        <span slot="message">
+          <span>
+            Slotted error message. And some slotted and deeply nested <a [href]="'#'">anchor</a>, <b>bold</b>,
+            <strong>strong</strong>, <em>emphasized</em> and <i>italic</i> text.
+          </span>
+        </span>
       </p-checkbox-wrapper>
     </div>
 
     <div class="playground dark" title="should render with slotted content with error state and message">
       <p-checkbox-wrapper [theme]="'dark'" [state]="'error'">
-        <span slot="label">Some label with a <a [href]="'https://designsystem.porsche.com'">link</a>.</span>
+        <span slot="label">
+          <span>
+            Slotted label. And some slotted and deeply nested <a [href]="'#'">anchor</a>, <b>bold</b>, <strong>strong</strong>,
+            <em>emphasized</em> and <i>italic</i> text.
+          </span>
+        </span>
         <input [type]="'checkbox'" [name]="'some-name'" />
-        <span slot="message">Some error message with a <a [href]="'https://designsystem.porsche.com'">link</a>.</span>
+        <span slot="message">
+          <span>
+            Slotted error message. And some slotted and deeply nested <a [href]="'#'">anchor</a>, <b>bold</b>,
+            <strong>strong</strong>, <em>emphasized</em> and <i>italic</i> text.
+          </span>
+        </span>
+      </p-checkbox-wrapper>
+    </div>
+
+    <div class="playground light" title="should render with slotted content with success state and message">
+      <p-checkbox-wrapper [state]="'success'">
+        <span slot="label">
+          <span>
+            Slotted label. And some slotted and deeply nested <a [href]="'#'">anchor</a>, <b>bold</b>, <strong>strong</strong>,
+            <em>emphasized</em> and <i>italic</i> text.
+          </span>
+        </span>
+        <input [type]="'checkbox'" [name]="'some-name'" />
+        <span slot="message">
+          <span>
+            Slotted success message. And some slotted and deeply nested <a [href]="'#'">anchor</a>, <b>bold</b>,
+            <strong>strong</strong>, <em>emphasized</em> and <i>italic</i> text.
+          </span>
+        </span>
+      </p-checkbox-wrapper>
+    </div>
+
+    <div class="playground dark" title="should render with slotted content with success state and message">
+      <p-checkbox-wrapper [theme]="'dark'" [state]="'success'">
+        <span slot="label">
+          <span>
+            Slotted label. And some slotted and deeply nested <a [href]="'#'">anchor</a>, <b>bold</b>, <strong>strong</strong>,
+            <em>emphasized</em> and <i>italic</i> text.
+          </span>
+        </span>
+        <input [type]="'checkbox'" [name]="'some-name'" />
+        <span slot="message">
+          <span>
+            Slotted success message. And some slotted and deeply nested <a [href]="'#'">anchor</a>, <b>bold</b>,
+            <strong>strong</strong>, <em>emphasized</em> and <i>italic</i> text.
+          </span>
+        </span>
       </p-checkbox-wrapper>
     </div>