+++ packages/components-react/src/pages/ButtonPure.tsx
@@ -2,66 +2,54 @@
 import { PButtonPure } from '@porsche-design-system/components-react';
 
 export const ButtonPurePage = (): JSX.Element => {
-  const style = `
-    p-button-pure:not(:last-child) {
-      margin-right: 16px;
-    }
-
-    .stretched-buttons p-button-pure {
-      margin-right: 0;
-    }
-  `;
-
   return (
     <>
-      <style dangerouslySetInnerHTML={{ __html: style }} />
-
-      <div className="playground light" title="should render button with label">
+      <div className="playground light auto-layout" title="should render button with label">
         <PButtonPure>Label default</PButtonPure>
         <PButtonPure disabled={true}>Label disabled</PButtonPure>
         <PButtonPure loading={true}>Label loading</PButtonPure>
       </div>
-      <div className="playground dark" title="should render button with label on dark theme">
+      <div className="playground dark auto-layout" title="should render button with label on dark theme">
         <PButtonPure theme="dark">Label default</PButtonPure>
         <PButtonPure disabled={true} theme="dark">Label disabled</PButtonPure>
         <PButtonPure loading={true} theme="dark">Label loading</PButtonPure>
       </div>
 
-      <div className="playground light" title="should render button without label">
+      <div className="playground light auto-layout" title="should render button without label">
         <PButtonPure hideLabel={true}>Some label</PButtonPure>
         <PButtonPure hideLabel={true} disabled={true}>Some label</PButtonPure>
         <PButtonPure hideLabel={true} loading={true}>Some label</PButtonPure>
       </div>
-      <div className="playground dark" title="should render button without label on dark theme">
+      <div className="playground dark auto-layout" title="should render button without label on dark theme">
         <PButtonPure hideLabel={true} theme="dark">Some label</PButtonPure>
         <PButtonPure hideLabel={true} disabled={true} theme="dark">Some label</PButtonPure>
         <PButtonPure hideLabel={true} loading={true} theme="dark">Some label</PButtonPure>
       </div>
 
-      <div className="playground light" title="should render button with responsive label">
+      <div className="playground light auto-layout" title="should render button with responsive label">
         <PButtonPure hideLabel={{ base: true, xs: false, s: true, m: false, l: true, xl: false }}>
           Label responsive
         </PButtonPure>
       </div>
 
-      <div className="playground light" title="should render button without icon">
+      <div className="playground light auto-layout" title="should render button without icon">
         <PButtonPure icon="none">Label default</PButtonPure>
         <PButtonPure icon="none" disabled={true}>Label disabled</PButtonPure>
         <PButtonPure icon="none" loading={true}>Label loading</PButtonPure>
       </div>
-      <div className="playground dark" title="should render button with label on dark theme">
+      <div className="playground dark auto-layout" title="should render button with label on dark theme">
         <PButtonPure icon="none" theme="dark">Label default</PButtonPure>
         <PButtonPure icon="none" disabled={true} theme="dark">Label disabled</PButtonPure>
         <PButtonPure icon="none" loading={true} theme="dark">Label loading</PButtonPure>
       </div>
 
-      <div className="playground light" title="should not render button with different weight">
+      <div className="playground light auto-layout" title="should not render button with different weight">
         <PButtonPure weight="regular">Label weight regular</PButtonPure>
         <PButtonPure weight="semi-bold">Label weight semi-bold</PButtonPure>
         <PButtonPure weight="bold">Label weight bold</PButtonPure>
       </div>
 
-      <div className="playground light" title="should render with active state">
+      <div className="playground light auto-layout" title="should render with active state">
         <PButtonPure active={true}>Label active</PButtonPure>
         <PButtonPure active={true} disabled={true}>Label active disabled</PButtonPure>
         <PButtonPure active={true} loading={true}>Label active loading</PButtonPure>
@@ -73,7 +61,7 @@ export const ButtonPurePage = (): JSX.Element => {
         <PButtonPure active={true} hideLabel={true} loading={true}>Label active loading</PButtonPure>
       </div>
 
-      <div className="playground dark" title="should render with active state on dark background">
+      <div className="playground dark auto-layout" title="should render with active state on dark background">
         <PButtonPure active={true} theme="dark">Label active</PButtonPure>
         <PButtonPure active={true} disabled={true} theme="dark">Label active disabled</PButtonPure>
         <PButtonPure active={true} loading={true} theme="dark">Label active loading</PButtonPure>
@@ -85,49 +73,49 @@ export const ButtonPurePage = (): JSX.Element => {
         <PButtonPure active={true} hideLabel={true} loading={true} theme="dark">Label active loading</PButtonPure>
       </div>
 
-      <div className="playground light" title="should render button with specific icon">
+      <div className="playground light auto-layout" title="should render button with specific icon">
         <PButtonPure icon="delete">Label with specific icon</PButtonPure>
         <PButtonPure iconSource="./assets/icon-custom-kaixin.svg">Label with local icon-source</PButtonPure>
       </div>
 
-      <div className="playground dark" title="should render button with specific icon on dark background">
+      <div className="playground dark auto-layout" title="should render button with specific icon on dark background">
         <PButtonPure icon="delete" theme="dark">Label with specific icon</PButtonPure>
         <PButtonPure iconSource="./assets/icon-custom-kaixin.svg" theme="dark"
           >Label with local icon-source
         </PButtonPure>
       </div>
 
-      <div className="playground light" title="should render button with multiline label">
+      <div className="playground light auto-layout" title="should render button with multiline label">
         <PButtonPure style={{ width: '15rem' }}>Label multiline lorem ipsum dolor sit amet, consetetur sadipscing</PButtonPure>
       </div>
 
-      <div className="playground light" title="should render button-pure with custom clickable area">
+      <div className="playground light auto-layout" title="should render button-pure with custom clickable area">
         <PButtonPure style={{ padding: '1rem' }}>Label with custom click-area</PButtonPure>
         <PButtonPure hideLabel={true} style={{ padding: '1rem' }}>Label with custom click-area</PButtonPure>
       </div>
 
-      <div className="playground light" title="should render with no icon">
+      <div className="playground light auto-layout" title="should render with no icon">
         <PButtonPure icon="none">Label icon none</PButtonPure>
       </div>
 
-      <div className="playground light" title="should render icon if hide-label and icon none is set">
+      <div className="playground light auto-layout" title="should render icon if hide-label and icon none is set">
         <PButtonPure hideLabel={true} icon="none">Label hide-label icon none</PButtonPure>
       </div>
 
-      <div className="playground light" title="should align label to the left">
+      <div className="playground light auto-layout" title="should align label to the left">
         <PButtonPure alignLabel="left">Label align left</PButtonPure>
       </div>
-      <div className="playground light" title="should align label to the left or right depending on viewport">
+      <div className="playground light auto-layout" title="should align label to the left or right depending on viewport">
         <PButtonPure alignLabel={{ base: 'left', xs: 'right', s: 'left', m: 'right', l: 'left', xl: 'right' }}>
           Label align responsive
         </PButtonPure>
       </div>
 
-      <div className="playground light stretched-buttons" title="should render with stretched label">
+      <div className="playground light auto-layout" title="should render with stretched label">
         <PButtonPure stretch={true}>Label stretch</PButtonPure>
         <PButtonPure stretch={true} alignLabel="left">Label stretch align left</PButtonPure>
       </div>
-      <div className="playground light" title="should render with stretched label depending on viewport">
+      <div className="playground light auto-layout" title="should render with stretched label depending on viewport">
         <PButtonPure stretch={{ base: true, xs: false, s: true, m: false, l: true, xl: false }}>
           Label stretch responsive
         </PButtonPure>
@@ -147,7 +135,7 @@ export const ButtonPurePage = (): JSX.Element => {
         <PButtonPure size="inherit" style={{ fontSize: '48px' }}>Label size inherit</PButtonPure>
       </div>
 
-      <div className="playground light" title="should render button with responsive size">
+      <div className="playground light auto-layout" title="should render button with responsive size">
         <PButtonPure
           size={{ base: 'x-small', xs: 'small', s: 'medium', m: 'large', l: 'x-large', xl: 'inherit' }}
           style={{ fontSize: '48px' }}
@@ -158,6 +146,7 @@ export const ButtonPurePage = (): JSX.Element => {
 
       <div className="playground light" title="should render with no icon and size inherit">
         <PButtonPure icon="none" size="inherit" style={{ fontSize: '48px' }}>Label icon none size inherit</PButtonPure>
+        <br />
         <PButtonPure
           icon="none"
           size={{ base: 'x-small', xs: 'small', s: 'medium', m: 'large', l: 'x-large', xl: 'inherit' }}