+++ packages/components/src/components/text-field-wrapper/text-field-wrapper-utils.ts
@@ -1,5 +1,5 @@
-import { getClosestHTMLElement, hasCounter } from '../../utils';
 import type { IconName } from '../../types';
+import { getClosestHTMLElement, hasCounter, hasDocument } from '../../utils';
 import { borderWidthBase } from '@porsche-design-system/utilities-v2';
 import { cssVariableInputPaddingLeft, cssVariableInputPaddingRight } from './text-field-wrapper-styles';
 
@@ -62,4 +62,19 @@ export const addInputEventListenerForSearch = (
 export const dispatchInputEvent = (el: HTMLInputElement): void => {
   // { bubbles: true } is crucial for react onChange callback getting invoked
   el.dispatchEvent(new Event('input', { bubbles: true }));
+};
+
+// eslint-disable-next-line no-underscore-dangle
+export const _hasShowPickerSupport = (): boolean => {
+  return (
+    hasDocument &&
+    'showPicker' in HTMLInputElement.prototype &&
+    // TODO: it would be better to determinate support by checking for existence of "calendar-picker-indicator"
+    !!window.navigator.userAgent.match(/chrome|chromium|crios|edg/i)
+  );
+};
+const hasShowPickerSupport = _hasShowPickerSupport();
+
+export const showCustomCalendarOrTimeIndicator = (isCalendar: boolean, isTime: boolean): boolean => {
+  return hasShowPickerSupport && (isCalendar || isTime);
 };