+++ packages/utilities/projects/utilities/src/js/theme/themeLightStateHover.ts
@@ -1,1 +1,1 @@
-export const themeLightStateHover = 'rgba(148, 149, 152, 0.20)';
+export const themeLightStateHover = 'rgba(148, 149, 152, .18)';